+++ server/aggregators/topNHistoricAvg/main.go
@@ -44,7 +44,7 @@ func getConfig() (config, error) {
 
 type handler struct {
 	topN    int
-	results utils.GameHeap
+	results middleware.GameHeap
 }
 
 func (h handler) Aggregate(_ *middleware.Channel, batch middleware.Batch[middleware.Game]) error {
@@ -99,7 +99,7 @@ func main() {
 
 	h := handler{
 		topN:    cfg.TopN,
-		results: make(utils.GameHeap, cfg.TopN),
+		results: make(middleware.GameHeap, cfg.TopN),
 	}
 	ctx, _ := signal.NotifyContext(context.Background(), syscall.SIGTERM)
 	agg, err := aggregator.NewAggregator(aggCfg, h)
